refactor: deduplicate post metadata validation
