+++ scripts/check-frontmatter.mjs
@@ -1,29 +1,10 @@
 import fs from 'node:fs/promises';
 import path from 'node:path';
 import { fileURLToPath } from 'node:url';
+import { isRealDate, parseFrontmatter, validatePostSourcePath } from './post-metadata.mjs';
 
 const scriptDir = path.dirname(fileURLToPath(import.meta.url));
 const root = path.resolve(scriptDir, '..');
-
-function parseFrontmatter(raw) {
-  if (!raw.startsWith('---\n')) return { data: {}, body: raw, hasFrontmatter: false };
-  const end = raw.indexOf('\n---\n', 4);
-  if (end === -1) return { data: {}, body: raw, hasFrontmatter: false };
-  const fm = raw.slice(4, end).trim();
-  const body = raw.slice(end + 5);
-  const data = {};
-  for (const line of fm.split('\n')) {
-    const idx = line.indexOf(':');
-    if (idx === -1) continue;
-    const key = line.slice(0, idx).trim();
-    let value = line.slice(idx + 1).trim();
-    if ((value.startsWith('"') && value.endsWith('"')) || (value.startsWith("'") && value.endsWith("'"))) {
-      value = value.slice(1, -1);
-    }
-    data[key] = value;
-  }
-  return { data, body, hasFrontmatter: true };
-}
 
 async function listMarkdownFiles(dir) {
   try {
@@ -35,20 +16,6 @@ async function listMarkdownFiles(dir) {
     if (error?.code === 'ENOENT') return [];
     throw error;
   }
-}
-
-function isIsoDate(value) {
-  return /^\d{4}-\d{2}-\d{2}$/.test(value);
-}
-
-function isRealDate(value) {
-  if (!isIsoDate(value)) return false;
-  const d = new Date(`${value}T00:00:00Z`);
-  return !Number.isNaN(d.getTime()) && d.toISOString().slice(0, 10) === value;
-}
-
-function expectedPostSlug(filename) {
-  return filename.replace(/\.md$/, '');
 }
 
 function validatePost(file, raw, errors) {
@@ -69,16 +36,10 @@ function validatePost(file, raw, errors) {
     errors.push(`${rel}: date must be a real ISO date (YYYY-MM-DD), got \`${data.date}\``);
   }
 
-  const expectedSlug = expectedPostSlug(name);
-  if (data.slug && data.slug !== expectedSlug) {
-    errors.push(`${rel}: slug must match filename \`${expectedSlug}\`, got \`${data.slug}\``);
-  }
-
-  const filenameDate = name.match(/^(\d{4}-\d{2}-\d{2})-/)?.[1];
-  if (!filenameDate) {
-    errors.push(`${rel}: filename must use YYYY-MM-DD-slug.md`);
-  } else if (data.date && data.date !== filenameDate) {
-    errors.push(`${rel}: date \`${data.date}\` does not match filename prefix \`${filenameDate}\``);
+  try {
+    validatePostSourcePath(name, data.date, data.slug);
+  } catch (error) {
+    errors.push(`${rel}: ${error.message}`);
   }
 
   if (!body.trim()) {
